@@ -4,28 +4,30 @@
   "private": true,
   "proxy": "http://localhost:3003",
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.28",
-    "@fortawesome/free-solid-svg-icons": "^5.13.0",
-    "@fortawesome/react-fontawesome": "^0.1.17",
-    "@xstate/react": "^2.0.0",
-    "chart.js": "^3.7.0",
-    "d3": "^7.1.0",
+    "@fortawesome/fontawesome-svg-core": "^6.4.0",
+    "@fortawesome/free-solid-svg-icons": "^6.4.0",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@xstate/react": "^3.2.2",
+    "chart.js": "^4.3.0",
+    "d3": "^7.8.4",
     "d3-cloud": "^1.2.5",
     "d3-selection": "^3.0.0",
-    "immer": "^9.0.12",
+    "immer": "^10.0.2",
     "randomcolor": "^0.6.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.2.1",
-    "styled-components": "^5.3.3",
-    "typescript": "^4.5.5",
-    "xstate": "^4.29.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.11.2",
+    "styled-components": "^6.0.0-rc.1",
+    "typescript": "^5.0.4",
+    "uuid": "^9.0.0",
+    "xstate": "^4.37.2"
   },
   "scripts": {
     "start": "vite",
     "build": "tsc && vite build",
     "serve": "vite preview",
     "test": "jest",
+    "watch": "jest --watch",
     "lint": "eslint --ext .js,.jsx,.ts,.tsx src",
     "check": "tsc && npm run lint",
     "lint-fix": "eslint --ext .js,.jsx,.ts,.tsx src --fix"
