@@ -1,5 +1,5 @@
 /* eslint-disable no-param-reassign */
-import produce from 'immer';
+import { produce } from 'immer';
 import { CloudState, CloudActions } from './cloudReducer.types';
 
 export const initialState = {
